requirements.txt
flask
websockets
asyncio
pyaudio
numpy==1.21.6
tensorflow==2.8.0
librosa
paho-mqtt